LayoutTests/svg/dynamic-updates/SVGFEConvolveMatrixElement-svgdom-order-prop-expected.txt

CONSOLE MESSAGE: line 42: TypeError: 'undefined' is not an object (evaluating 'convolveMatrixElement.orderX.baseVal = 3')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'order' property of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL convolveMatrixElement.orderX.baseVal should be 2. Threw exception TypeError: 'undefined' is not an object (evaluating 'convolveMatrixElement.orderX.baseVal')
FAIL convolveMatrixElement.orderY.baseVal should be 2. Threw exception TypeError: 'undefined' is not an object (evaluating 'convolveMatrixElement.orderY.baseVal')
